fix performance, moved screens folder, updated collectibles on isDeveloperMode
