+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/collectibleManagement/index.tsx
@@ -1,6 +1,1 @@
-import { CollectibleManagementScreen } from 'features/portfolio/collectibles/components/CollectibleManagementScreen'
-import React from 'react'
-
-export default (): JSX.Element => {
-  return <CollectibleManagementScreen />
-}
+export { CollectibleManagementScreen as default } from 'features/portfolio/collectibles/screens/CollectibleManagementScreen'
